lookupLedgerReporting

+++ src/ripple/app/misc/impl/Transaction.cpp
@@ -144,7 +144,7 @@ Transaction::getLedgerSeq(uint256 const& id, Application& app)
         PQresultStatus(res.get()) == PGRES_TUPLES_OK ||
         PQresultStatus(res.get()) == PGRES_SINGLE_TUPLE);
     if (PQgetisnull(res.get(), 0, 0))
-        return 0;
+        return uint32_t(0);
 
     char const* resultStr = PQgetvalue(res.get(), 0, 0);
 
@@ -169,7 +169,7 @@ Transaction::getLedgerSeq(uint256 const& id, Application& app)
     }
     else
     {
-        return 0;
+        return uint32_t(0);
     }
 }
